Extract automation task execution into a dedicated module
